Move options related types that are linked into CodeStyle layer to a different namespace to avoid type name conflicts
