+++ src/CodeStyle/Directory.Build.targets
@@ -1,3 +1,7 @@
 <Project>
   <Import Project="..\..\Directory.Build.targets"/>
+
+  <ItemGroup>
+    <AdditionalFiles Include="$(MSBuildThisFileDirectory)\BannedSymbols.txt" />
+  </ItemGroup>
 </Project>
