+++ src/Features/Core/Portable/Diagnostics/AnalyzerHelper.cs
@@ -13,8 +13,6 @@ using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Diagnostics.Log;
 using Microsoft.CodeAnalysis.Diagnostics.Telemetry;
 using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Internal.Log;
@@ -24,7 +22,7 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Diagnostics
 {
-    internal static class AnalyzerHelper
+    internal static partial class AnalyzerHelper
     {
         private const string CSharpCompilerAnalyzerTypeName = "Microsoft.CodeAnalysis.Diagnostics.CSharp.CSharpCompilerDiagnosticAnalyzer";
         private const string VisualBasicCompilerAnalyzerTypeName = "Microsoft.CodeAnalysis.Diagnostics.VisualBasic.VisualBasicCompilerDiagnosticAnalyzer";
@@ -109,92 +107,19 @@ namespace Microsoft.CodeAnalysis.Diagnostics
             return new AnalyzerConfigOptionSet(configOptions, optionSet);
         }
 
-        public static T GetOption<T>(this SemanticModelAnalysisContext context, Option<T> option)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.SemanticModel.SyntaxTree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, syntaxTree, cancellationToken);
-        }
-
-        public static T GetOption<T>(this SyntaxNodeAnalysisContext context, Option<T> option)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.Node.SyntaxTree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, syntaxTree, cancellationToken);
-        }
-
-        public static T GetOption<T>(this SyntaxTreeAnalysisContext context, Option<T> option)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.Tree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, syntaxTree, cancellationToken);
-        }
-
-        public static T GetOption<T>(this OperationAnalysisContext context, Option<T> option)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.Operation.Syntax.SyntaxTree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, syntaxTree, cancellationToken);
-        }
-
         public static T GetOption<T>(this AnalyzerOptions analyzerOptions, Option<T> option, SyntaxTree syntaxTree, CancellationToken cancellationToken)
         {
             return GetOptionAsync<T>(analyzerOptions, option, language: null, syntaxTree, cancellationToken).GetAwaiter().GetResult();
         }
 
-        public static T GetOption<T>(this SemanticModelAnalysisContext context, PerLanguageOption<T> option, string language)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.SemanticModel.SyntaxTree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, language, syntaxTree, cancellationToken);
-        }
-
-        public static T GetOption<T>(this SyntaxNodeAnalysisContext context, PerLanguageOption<T> option, string language)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.Node.SyntaxTree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, language, syntaxTree, cancellationToken);
-        }
-
-        public static T GetOption<T>(this SyntaxTreeAnalysisContext context, PerLanguageOption<T> option, string language)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.Tree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, language, syntaxTree, cancellationToken);
-        }
-
-        public static T GetOption<T>(this OperationAnalysisContext context, PerLanguageOption<T> option, string language)
-        {
-            var analyzerOptions = context.Options;
-            var syntaxTree = context.Operation.Syntax.SyntaxTree;
-            var cancellationToken = context.CancellationToken;
-
-            return GetOption(analyzerOptions, option, language, syntaxTree, cancellationToken);
-        }
-
-        public static T GetOption<T>(this AnalyzerOptions analyzerOptions, PerLanguageOption<T> option, string language, SyntaxTree syntaxTree, CancellationToken cancellationToken)
+        public static T GetOption<T>(this AnalyzerOptions analyzerOptions, PerLanguageOption<T> option, string? language, SyntaxTree syntaxTree, CancellationToken cancellationToken)
         {
             return GetOptionAsync<T>(analyzerOptions, option, language, syntaxTree, cancellationToken).GetAwaiter().GetResult();
         }
 
         public static async ValueTask<T> GetOptionAsync<T>(this AnalyzerOptions analyzerOptions, IOption option, string? language, SyntaxTree syntaxTree, CancellationToken cancellationToken)
         {
-            var configOptions = analyzerOptions.AnalyzerConfigOptionsProvider.GetOptions(syntaxTree);
-            if (configOptions.TryGetEditorConfigOption<T>(option, out var value))
+            if (analyzerOptions.TryGetEditorConfigOption<T>(option, syntaxTree, out var value))
             {
                 return value;
             }
@@ -204,11 +129,6 @@ namespace Microsoft.CodeAnalysis.Diagnostics
 #pragma warning restore CS0612 // Type or member is obsolete
             return (T)optionSet?.GetOption(new OptionKey(option, language)) ?? (T)option.DefaultValue!;
         }
-
-        [Obsolete]
-        [PerformanceSensitive("https://github.com/dotnet/roslyn/issues/23582", OftenCompletesSynchronously = true)]
-        public static OptionSet? GetOptions(this AnalyzerOptions analyzerOptions, SyntaxTree syntaxTree, CancellationToken cancellationToken)
-            => analyzerOptions.GetDocumentOptionSetAsync(syntaxTree, cancellationToken).GetAwaiter().GetResult();
 
         [PerformanceSensitive("https://github.com/dotnet/roslyn/issues/23582", OftenCompletesSynchronously = true)]
         public static ValueTask<OptionSet?> GetDocumentOptionSetAsync(this AnalyzerOptions analyzerOptions, SyntaxTree syntaxTree, CancellationToken cancellationToken)
