+++ src/Workspaces/Core/Portable/CodeStyle/CodeStyleOption.cs
@@ -7,7 +7,11 @@ using System.Collections.Generic;
 using System.Xml.Linq;
 using Microsoft.CodeAnalysis.Diagnostics;
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.CodeStyle
+#endif
 {
     internal interface ICodeStyleOption
     {
