+++ src/Workspaces/Core/Portable/Options/EditorConfig/EditorConfigStorageLocationExtensions.cs
@@ -4,7 +4,11 @@ using System;
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis.Diagnostics;
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     internal static class EditorConfigStorageLocationExtensions
     {
