+++ src/Workspaces/Core/Portable/Options/EditorConfig/EditorConfigStorageLocation`1.cs
@@ -14,7 +14,11 @@ using System.Linq;
 using OptionSet = Microsoft.CodeAnalysis.Diagnostics.AnalyzerConfigOptions;
 #endif
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     /// <summary>
     /// Specifies that an option should be read from an .editorconfig file.
