+++ src/Workspaces/Core/Portable/Options/EditorConfig/IEditorConfigStorageLocation2.cs
@@ -8,7 +8,11 @@
 using OptionSet = Microsoft.CodeAnalysis.Diagnostics.AnalyzerConfigOptions;
 #endif
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     internal interface IEditorConfigStorageLocation2 : IEditorConfigStorageLocation
     {
