+++ src/Workspaces/Core/Portable/Options/IOption.cs
@@ -7,7 +7,11 @@
 using System;
 using System.Collections.Immutable;
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     public interface IOption
     {
