+++ src/Workspaces/Core/Portable/Options/IOptionWithGroup.cs
@@ -4,7 +4,11 @@
 
 #nullable enable
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     /// <summary>
     /// Group/sub-feature associated with an <see cref="IOption"/>.
