+++ src/Workspaces/Core/Portable/Options/OptionGroup.cs
@@ -6,7 +6,11 @@
 
 using System;
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     /// <summary>
     /// Group/sub-feature associated with an <see cref="IOption"/>.
