+++ src/Workspaces/Core/Portable/Options/OptionKey.cs
@@ -11,7 +11,11 @@ using Roslyn.Utilities;
 using WorkspacesResources = Microsoft.CodeAnalysis.CodeStyleResources;
 #endif
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     [NonDefaultable]
     public readonly struct OptionKey : IEquatable<OptionKey>
