+++ src/Workspaces/Core/Portable/Options/OptionStorageLocation.cs
@@ -4,7 +4,11 @@
 
 #nullable enable
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     /// <summary>
     /// The base type of all types that specify where options are stored.
