+++ src/Workspaces/Core/Portable/Options/Option`1.cs
@@ -7,7 +7,11 @@
 using System;
 using System.Collections.Immutable;
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.Internal.Options
+#else
 namespace Microsoft.CodeAnalysis.Options
+#endif
 {
     /// <summary>
     /// Marker interface for <see cref="Option{T}"/>
