+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/CodeStyle/TypeStyle/UseVarPreference.cs
@@ -4,7 +4,11 @@
 
 using System;
 
+#if CODE_STYLE
+namespace Microsoft.CodeAnalysis.CSharp.Internal.CodeStyle.TypeStyle
+#else
 namespace Microsoft.CodeAnalysis.CSharp.CodeStyle.TypeStyle
+#endif
 {
     [Flags]
     internal enum UseVarPreference
