+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/Utilities/TypeStyle/CSharpUseImplicitTypeHelper.cs
@@ -6,7 +6,6 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.CodeStyle.TypeStyle;
 using Microsoft.CodeAnalysis.CSharp.Extensions;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Shared.Extensions;
@@ -15,8 +14,10 @@ using Microsoft.CodeAnalysis.Simplification;
 
 #if CODE_STYLE
 using OptionSet = Microsoft.CodeAnalysis.Diagnostics.AnalyzerConfigOptions;
+using Microsoft.CodeAnalysis.CSharp.Internal.CodeStyle.TypeStyle;
 #else
 using Microsoft.CodeAnalysis.Options;
+using Microsoft.CodeAnalysis.CSharp.CodeStyle.TypeStyle;
 #endif
 
 namespace Microsoft.CodeAnalysis.CSharp.Utilities
