+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Formatting/FormattingOptions.cs
@@ -6,11 +6,13 @@
 
 using System;
 using System.Collections.Immutable;
-using Microsoft.CodeAnalysis.Options;
 using Roslyn.Utilities;
 
 #if CODE_STYLE
+using Microsoft.CodeAnalysis.Internal.Options;
 using WorkspacesResources = Microsoft.CodeAnalysis.CodeStyleResources;
+#else
+using Microsoft.CodeAnalysis.Options;
 #endif
 
 namespace Microsoft.CodeAnalysis.Formatting
